Move train_detector_from_votes to tests/helpers.py
conftest.py cannot be imported as a regular module. Move the shared
helper to tests/helpers.py which is importable from test files.
