+++ tests/test_cli_autodetect.py
@@ -10,7 +10,7 @@ import numpy as np
 import pytest
 
 import app as app_module
-from conftest import train_detector_from_votes
+from helpers import train_detector_from_votes
 from vtsearch.cli import (
     _build_multi_results_dict,
     _detect_media_type,
