+++ tests/test_csv_webhook_exporters.py
@@ -9,7 +9,7 @@ from unittest import mock
 import pytest
 
 import app as app_module
-from conftest import train_detector_from_votes
+from helpers import train_detector_from_votes
 from vtsearch.cli import (
     _run_exporter,
     run_autodetect,
