+++ tests/test_detectors.py
@@ -5,7 +5,7 @@ import numpy as np
 import pytest
 
 import app as app_module
-from conftest import train_detector_from_votes
+from helpers import train_detector_from_votes
 
 
 class TestDetectorSort:
