+++ tests/test_integration.py
@@ -18,7 +18,7 @@ import numpy as np
 import pytest
 
 import app as app_module
-from conftest import train_detector_from_votes
+from helpers import train_detector_from_votes
 
 
 # ---------------------------------------------------------------------------
